Edit corresponding java classes that are affected from moving TaskList class to duke.task
