+++ data/duke.txt
@@ -1,1 +1,3 @@
 T | false | sleep
+T | true | feed tofu
+D | false | sleep | 2024-01-01
