+++ src/main/java/duke/Storage.java
@@ -1,6 +1,8 @@
 package duke;
 
 import duke.task.Task;
+import duke.task.TaskList;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.IOException;
