+++ src/main/java/duke/command/ExitCommand.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.DukeException;
-import duke.TaskList;
+import duke.task.TaskList;
 import duke.ui.Ui;
 
 public class ExitCommand implements Command {
