+++ src/test/java/duke/TaskListTest.java
@@ -2,8 +2,9 @@ package duke;
 
 import duke.task.Deadline;
 import duke.task.Event;
+import duke.task.TaskList;
 import duke.task.ToDo;
-import org.junit.jupiter.api.Test;
+
 import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
